% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{concordia_download_case}
\alias{concordia_download_case}
\title{Download Specific Case Data in JSON Format}
\usage{
concordia_download_case(id, path)
}
\arguments{
\item{id}{A numeric or character value that specifies the id of the case to be downloaded.}

\item{path}{A character string specifying the path where the downloaded data should be saved.}
}
\value{
This function does not explicitly return a value. Instead, it writes the downloaded data to a file at the specified path.
}
\description{
This function checks if the JSON file for a specific case exists at the given path.
If the file does not exist, it sends a GET request to a specific URL to download the case data in JSON format.
}
