% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{concordia_download_doc}
\alias{concordia_download_doc}
\title{Download Specific Case Data in PDF/DOC/DOCX Format}
\usage{
concordia_download_doc(data_doc, path)
}
\arguments{
\item{data_doc}{A data frame with pdf information(id of the case, id of the file, name of the file)}

\item{path}{A character string specifying the path where the downloaded data should be saved.}
}
\value{
This function does not explicitly return a value. Instead, it writes the downloaded data to a file at the specified path.
}
\description{
This function checks if the PDF/DOC/DOCX file for a specific case exists at the given path.
If the file does not exist, it sends a GET request to a specific URL to download the case data in PDF/DOC/DOCX format.
}
