% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{concordia_download_pag}
\alias{concordia_download_pag}
\title{Download Data from a Specific Page}
\usage{
concordia_download_pag(pag, path)
}
\arguments{
\item{pag}{A numeric or character value that specifies the page number from which the data should be downloaded.}

\item{path}{A character string specifying the path where the downloaded data should be saved.}
}
\value{
This function does not explicitly return a value.
Instead, it writes the downloaded data to a file at the specified path.
}
\description{
This function sends a POST request to a specific URL to download data from a given page.
The downloaded data is cleaned and transformed into a tibble data structure.
The tibble is then written to a file and saved to a specified path.
}
