% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{concordia_pages}
\alias{concordia_pages}
\title{Itamaraty API Cases Dataset}
\format{
A data frame with 12,156 rows and 16 variables:
\describe{
\item{id}{A character vector representing the ID of the case.}
\item{page}{An integer vector indicating the page number from which the case data was downloaded.}
\item{count}{An integer vector indicating the count of cases present in each page.}
\item{total_pages}{An integer vector indicating the total number of pages of cases data.}
\item{total_count}{An integer vector indicating the total count of cases.}
\item{id_2}{A character vector representing another ID associated with the case.}
\item{id_3}{An integer vector representing yet another ID associated with the case.}
\item{titulo}{A character vector representing the title or description of the case.}
\item{assuntos}{A list of data frames, each containing detailed information about the case.}
\item{vigencia}{A character vector indicating the current status of the case.}
\item{outra_parte}{A character vector indicating the other party involved in the case.}
\item{data_celebracao}{A character vector indicating the date of celebration or conclusion of the case.}
\item{outras_partes}{A list of character vectors, each indicating the other parties involved in the case.}
\item{tipo_acordo}{A character vector indicating the type of agreement involved in the case.}
\item{documentos}{A list of data frames, each containing documents related to the case.}
\item{documento_integra}{A data frame containing detailed documents related to the case.}
}
}
\source{
\url{https://concordia.itamaraty.gov.br/}
}
\usage{
concordia_pages
}
\description{
A dataset containing 12,156 cases retrieved from the Itamaraty API.
}
\keyword{datasets}
